validation

+++ src/app/components/employee-details/employee-details.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Employee } from 'src/app/model/employee.model';
 import { EmployeeService } from 'src/app/services/employee.service';
+import { FlashMessagesService } from 'angular2-flash-messages';
 
 
 @Component({
@@ -14,9 +15,10 @@ export class EmployeeDetailsComponent implements OnInit {
   name: '';
   id: number;
   employee: Employee;
+  showMsg: boolean=false;
 
   constructor(private route: ActivatedRoute,private router: Router,
-    private employeeService: EmployeeService) { }
+    private employeeService: EmployeeService, private flashMessage: FlashMessagesService) { }
 
   ngOnInit() {
     this.employee = new Employee();
@@ -47,13 +49,25 @@ export class EmployeeDetailsComponent implements OnInit {
     //   });
 
     // }
-    this.employeeService.deleteEmployee(id)
+    if(confirm("Are you sure to delete " + this.employee.empName + " data ?")){
+      console.log("Implement delete functionality here");
+      this.employeeService.deleteEmployee(id)
       .subscribe(
         data => {
           console.log(data);
-          this.list();
+          //this.showMsg =true;
+          this.showFlash();
+          //this.list();
         },
         error => console.log(error));
+    }
+  }
+
+  showFlash() {
+    // 1st parameter is a flash message text
+    // 2nd parameter is optional. You can pass object with options.
+    this.flashMessage.show('Deleted Successfully', { cssClass: 'custom-success', timeout: 3000 });
+    
   }
 
   empUpdate(id: number){
